Upload_docs/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
albucore==0.0.24
albumentations==2.0.8
altair==5.3.0
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
anykeystore==0.2
apex==0.9.10.dev0
appnope==0.1.4
asgiref==3.10.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
audioread==3.0.1
autograd==1.8.0
autograd-gamma==0.5.0
banks==2.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.5
Cartopy==0.23.0
cattrs==23.2.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
controlnet-aux==0.0.9
coremltools==7.1
coverage==5.5
cryptacular==1.6.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dearpygui==1.10.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diffusers==0.32.2
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
Django==4.2.26
dnspython==2.7.0
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
eventlet==0.38.0
exceptiongroup==1.2.0
executing==2.1.0
eyed3==0.9.7
fastapi==0.128.0
filelock==3.13.1
filetype==1.2.0
Flask==3.1.0
Flask-SocketIO==5.4.1
flatbuffers==23.5.26
fonttools==4.51.0
formulaic==1.2.0
frozenlist==1.8.0
fsspec==2024.2.0
ftfy==6.3.1
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.5.4
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
google_search_results==2.4.2
googlemaps==4.10.0
googlesearch-python==1.3.0
greenlet==3.1.1
griffe==1.14.0
grpcio==1.63.2
grpcio-tools==1.63.2
h11==0.14.0
h2==4.3.0
h5py==3.10.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
hupper==1.12.1
hyperframe==6.1.0
idna==3.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==7.0.1
importlib_resources==6.4.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.12.0
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
librosa==0.10.1
lifelines==0.30.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.13
llama-index-core==0.14.13
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.15
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.13.1
llama-parse==0.6.54
llvmlite==0.42.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
midas==0.7.0
ml-dtypes==0.2.0
mlxtend==0.23.2
mmh3==5.2.0
monai==1.5.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.7.1
mutagen==1.47.0
mypy_extensions==1.1.0
namex==0.0.7
narwhals==2.4.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nltk==3.9.2
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
onnx-coreml==1.3
openai==2.16.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.10.0
outcome==1.3.0.post0
packaging==23.2
paho-mqtt==2.1.0
pandas==2.2.1
param==2.2.0
parso==0.8.4
PasteDeploy==3.1.0
patsy==0.5.6
pbkdf2==1.3
pexpect==4.9.0
pillow==10.2.0
pipenv==2024.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.4.0
pooch==1.8.1
postgrest==2.27.2
prince==0.13.0
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.4.1
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaml==23.12.0
pyarrow==18.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.5.2
Pygments==2.17.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymodbus==3.8.6
pyogrio==0.10.0
pyparsing==3.1.2
pypdf==6.6.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
pyramid==2.0.2
pyramid-mailer==0.15.1
pyreadstat==1.2.8
pyroaring==1.0.3
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
realtime==2.27.2
referencing==0.35.1
regex==2024.11.6
replicate==1.0.4
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
selenium==4.29.0
shapely==2.0.7
simple-websocket==1.1.0
simpleitk==2.5.2
simsimd==6.5.1
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
stability-sdk==0.8.6
stack-data==0.6.3
starlette==0.49.3
statsmodels==0.14.4
storage3==2.27.2
streamlit==1.40.1
StrEnum==0.4.15
strictyaml==1.7.3
stringzilla==3.12.6
stripe==11.6.0
striprtf==0.0.26
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
Theano==1.0.5
threadpoolctl==3.4.0
tifffile==2024.8.30
tiktoken==0.12.0
timm==1.0.16
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torchaudio==2.6.0
torchcam==0.4.0
torchgeometry==0.1.2
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transaction==5.0
transformers==4.50.3
translationstring==1.4
trio==0.29.0
trio-websocket==0.12.2
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.39.0
velruse==1.1.1
venusian==3.1.1
virtualenv==20.26.3
wcwidth==0.2.13
webdriver-manager==4.0.2
WebOb==1.8.9
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xgboost==2.1.3
yarl==1.22.0
yellowbrick==1.5
zipp==3.17.0
zmq==0.0.0
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
